requirements.txt
flask==2.2.5
firebase-admin==6.2.0
numpy<1.22
scipy<1.8
scikit-learn<1.1
opencv-python-headless
tensorflow-cpu==2.11.0
requests
mtcnn